#33 GameEngine refactoring
